makeWinInstaller.py starts to support unicode. Bugfixes for vplants_conf.py.
git-svn-id: https://scm.gforge.inria.fr/svn/openalea/branches/release_1_0@3444 04ea39c6-a500-0410-8508-e77fe2a861cf

+++ deploy/scripts/winInstallers/makeWinInstaller.py
@@ -19,7 +19,7 @@
 import argparse
 from collections import OrderedDict
 import glob
-from os.path import exists, join as pj, basename, abspath, split, dirname
+from os.path import exists, join as pj, basename, abspath, split, dirname, splitext
 import os
 import platform
 import shutil
@@ -33,11 +33,14 @@ from openalea.deploy.gforge_util import add_private_gforge_repositories
 from openalea.deploy.util import get_repo_list
 from distutils import log
 
+import locale
+unused, local_enc = locale.getdefaultlocale()
+
 pi = PackageIndex(search_path=[])
 pi.add_find_links(get_repo_list())
 
 err = sys.stderr.write
-out = sys.stdout.write
+
 __path__ = dirname(abspath(__file__))
 
 # -- Some Flags : DONT CHANGE THESE IF YOU DON'T KNOW WHAT YOU'RE DOING --
@@ -46,20 +49,21 @@ NOFLAG = 0
 EGG     = 2**0
 EXE     = 2**1
 ZIPDIST = 2**2
-MSI     = 2**3
+EXEDIST = 2**3
+MSI     = 2**4
 # Does the egg depend on python version?
-PY_DEP  = 2**4
+PY_DEP  = 2**5
 # Is it for runtime or development, or both ?
-RUNTIME = 2**5
-DEVELOP = 2**6
+RUNTIME = 2**6
+DEVELOP = 2**7
 # Should we test for it ?
-TEST_ME = 2**7
+TEST_ME = 2**8
 # If it can't be installed
-NOT_INSTALLABLE = 2**8 #if tested and not installable, then fatal!
+NOT_INSTALLABLE = 2**9 #if tested and not installable, then fatal!
 # Is it architecture dependent (i386 vs x86_64)?
-ARCH    = 2**9
+ARCH    = 2**10
 # A flag that marks the thing as an innosetup component
-COMPONENT = 2**10
+COMPONENT = 2**11
 
 
 # -- Installer and dependency packages description are declared here but are actually defined in the
@@ -79,6 +83,7 @@ def bt(val, bit):
 installerExtensions = { 0:"",
                         EGG:".egg",
                         EXE:".exe",
+                        EXEDIST:".exe",
                         ZIPDIST:".zip",
                         MSI:".msi",
                         }
@@ -149,26 +154,35 @@ python_package_ti_template_msi    = python_package_test_template+python_package_
 
 
 
-def prepare_working_dir(instDir):    
+def prepare_working_dir(instDir, no_del=False):
     if exists(instDir):
+        if no_del:
+            return
         print instDir, "will be deleted"
         shutil.rmtree(instDir, ignore_errors=False)
     print instDir, "will be created"
     os.makedirs(instDir)
 
 import traceback
-def find_installer_files(outDir, srcDir, pyMaj, pyMin, arch, dependencies):
+def find_installer_files(outDir, tpp_eggDir, srcDir, pyMaj, pyMin, arch, dependencies):
 
     arch = "win32" if arch == "x86" else "win64"
     
     def globInstaller(pk, mask):
-        
+        dir_ = srcDir
         identifier = pk+"*"
         if bt(mask, PY_DEP):
             identifier+=pyMaj+"."+pyMin+"*"
         if bt(mask, MSI): identifier+=".msi"
-        elif bt(mask, ZIPDIST): identifier+=".zip"
-        elif bt(mask, EGG): identifier+=".egg"
+        elif bt(mask, EXEDIST): 
+            identifier+=".exe"
+            dir_ = tpp_eggDir
+        elif bt(mask, ZIPDIST): 
+            identifier+=".zip"
+            dir_ = tpp_eggDir
+        elif bt(mask, EGG): 
+            identifier+=".egg"
+            dir_ = tpp_eggDir
         elif bt(mask, EXE): identifier+=".exe"
         else:
             raise Exception("Unknown installer type: " + pk +":"+str(mask))
@@ -176,28 +190,28 @@ def find_installer_files(outDir, srcDir, pyMaj, pyMin, arch, dependencies):
         try:
             if bt(mask, ARCH): #WE CARE ABOUT THE ARCH
                 if arch=="win32": #either it has 32 or nothing but not 64
-                    files = [f for f in glob.iglob(pj(srcDir, identifier))  if arch in f or ("win32" not in f and "64" not in f)]
+                    files = [f for f in glob.iglob(pj(dir_, identifier))  if arch in f or ("win32" not in f and "64" not in f)]
                 else:
-                    files = [f for f in glob.iglob(pj(srcDir, identifier))  if arch in f]
+                    files = [f for f in glob.iglob(pj(dir_, identifier))  if arch in f]
                 return sorted(files, lambda x, y: cmp(len(x), len(y)))[0]
             else:
-                return glob.glob(pj(srcDir, identifier))[0]
+                return glob.glob(pj(dir_, identifier))[0]
         except:
             #traceback.print_exc()
-            err("\tNo installer found for "+pk+" with for "+srcDir+" "+identifier+"\n")
+            err(u"\tNo installer found for %s for %s : %s\n"%(pk, dir_,identifier))
             return None
                         
     print "Gathering paths to binaries..."
     ok = True
     for pk, info in dependencies.iteritems():
         mask = info[0]
-        if not bt(mask, NOT_INSTALLABLE) and info[1] is not None:
+        if not bt(mask, NOT_INSTALLABLE) and info[1] is None:
             ef = globInstaller(pk, mask)
             if ef is None:
                 ok = False
                 continue
             info[1] = ef
-            out("\tWill install %s\n"%ef)        
+            print "\tWill install %s"%ef        
     return ok
 
 
@@ -224,7 +238,7 @@ def get_project_eggs(arch, globs, outDir, srcDir):
     localFiles = map(basename, files)
     for f, filename in zip(files, localFiles):
         egg_paths.append(f)
-        out("\tWill install %s\n"%f)
+        print "\tWill install %s"%f
     return egg_paths
 
 
@@ -236,7 +250,7 @@ def generate_inno_installer_setup_group(setup):
         basev = basename(v)
         if "file" in k.lower():
             src = abspath(v)
-            out("\t"+src+"\n")
+            print "\t"+src
         final += k + "=" + src + "\n"
     return final
     
@@ -272,6 +286,7 @@ def generate_pascal_test_install_code(dependencies):
             else:            
                 #"ti" stands for "test and install"
                 if bt(mask, MSI): template = python_package_ti_template_msi
+                elif bt(mask, EXEDIST): template = python_package_ti_template_zipdist
                 elif bt(mask, ZIPDIST): template = python_package_ti_template_zipdist
                 elif bt(mask, EGG): template = python_package_ti_template_egg
                 elif bt(mask, EXE): template = python_package_ti_template_exe
@@ -285,6 +300,7 @@ def generate_pascal_test_install_code(dependencies):
                 continue
             else:
                 if bt(mask, MSI): template = python_package_install_template_msi
+                elif bt(mask, EXEDIST): template = python_package_install_template_zipdist
                 elif bt(mask, ZIPDIST): template = python_package_install_template_zipdist
                 elif bt(mask, EGG): template = python_package_install_template_egg
                 elif bt(mask, EXE): template = python_package_install_template_exe
@@ -389,7 +405,6 @@ def configure_inno_setup(appname, appversion, dependencies, args, funcs, egg_pth
     f.close()
 
     template = StrictTemplate(s)
-    #eggs = options["eggs"]
     eggnum = len(egg_pths)
 
     eggArrayInit = ""
@@ -432,11 +447,12 @@ def configure_inno_setup(appname, appversion, dependencies, args, funcs, egg_pth
                             POSTINSTALLCODE=funcs["generate_pascal_post_install_code"](egg_pths),                            
                             )
 
-    
-    f = open( pj(args.outDir, appname+"_installer_"+modeStr+".iss"), "w" )
-    f.write(s)
+    fpath = pj(args.outDir, appname+"_installer_"+modeStr+".iss")
+    f = open( fpath, "w" )
+    f.write(s.encode(local_enc))
     f.close()
     print "ok"
+    return fpath
 
 def make_stitcher( eggDir, pyMaj, pyMin):
     """Creates a function that inserts the pyX.X string
@@ -445,7 +461,7 @@ def make_stitcher( eggDir, pyMaj, pyMin):
     def __stitch_egg_names(eggName):
         if pyfix in eggName:
             return pj(eggDir, eggName)
-        part = eggName.partition(".egg")
+        part = splitext(eggName)
         return pj(eggDir, part[0] + pyfix + part[1])
     return __stitch_egg_names
 
@@ -455,7 +471,7 @@ def processInstaller(mask, runtimeMode):
     return False                                   
    
 def download_egg(eggname, dir_):
-    out("Downloading %s\n"%eggname)
+    print "Downloading %s"%eggname
     return pi.download(eggname, dir_)
     
    
@@ -492,7 +508,10 @@ should NOT write "Openalea*py2.6.egg" because it will be done automatically.
     return msg
 
 def _parse_dict_string(d_str):
-    return eval(d_str) if not isinstance(d_str, dict) else d_str    
+    return eval(d_str) if not isinstance(d_str, dict) else d_str
+
+def _parse_unicode_abspath(path):
+    return abspath(unicode(path, encoding=local_enc))
     
 def parse_arguments():
     parser = argparse.ArgumentParser(description="Create InnoSetup installers for Windows, for Openalea, VPlants and Alinea",
@@ -503,15 +522,16 @@ def parse_arguments():
     parser.add_argument("--pyMaj", "-P", default=default, help="Major python version (default = %s)."%default, type=str)
     default = str(sys.version_info.minor)
     parser.add_argument("--pyMin", "-p", default=default, help="Minor python version (default = %s)."%default, type=str)
-    default = pj(abspath(os.curdir), "eggs", "thirdparty")
-    parser.add_argument("--srcDir", "-s", default=default, help="Directory to look for third party installers or eggs (default = %s)."%default, type=abspath)
-    default = pj(abspath(os.curdir), "output", "PROJECT")
-    parser.add_argument("--outDir", "-o", default=default, help="Directory to put output (default=%s)."%default, type=abspath)
+    default = pj(abspath(os.getcwd()), "eggs", "thirdparty")
+    parser.add_argument("--srcDir", "-s", default=default, help=u"Directory to look for third party installers or eggs (default = %s)."%default, type=_parse_unicode_abspath)
+    default = pj(abspath(os.getcwdu()), "output", "PROJECT")
+    parser.add_argument("--outDir", "-o", default=default, help=u"Directory to put output (default=%s)."%default, type=abspath)
     
     parser.add_argument("--eggGlobs", "-b", default=None, help="Pattern to match the PROJECT egg names (is not defined in CONFFILE). Use '|' to specify many patterns.")
     
-    default = pj(abspath(os.curdir), "eggs", "PROJECT")
-    parser.add_argument("--eggDir", "-e", default=default, help="Directory where we will look for the PROJECT eggs (default = %s)"%default, type=abspath)
+    default = pj(abspath(os.getcwdu()), "eggs", "PROJECT")
+    parser.add_argument("--eggDir", "-e", default=default, help=u"Directory where we will look for the PROJECT eggs (default = %s)"%default, type=abspath)
+    parser.add_argument("--tpp-eggDir", "-t", default=default, help="Directory where we will look for the third party eggs (default = srcDir)", type=abspath)
     
     parser.add_argument("--devel", "-d", action="store_const", const=False, default=True, help="Build Development Toolkit or Runtime (default=runtime)", dest="runtime")
     parser.add_argument("--fetch-online", "-f", action="store_const", const=True, default=False, help="Download eggs from online repositories and use them.")
@@ -542,8 +562,10 @@ def main():
         args.outDir = args.outDir.replace("PROJECT", args.project+"_"+sys.platform+"_"+args.pyMaj+"."+args.pyMin)
     else:
         args.outDir = pj(args.outDir, args.project+"_"+sys.platform+"_"+args.pyMaj+"."+args.pyMin)
+    args.tpp_eggDir = args.tpp_eggDir or args.srcDir
+    print args.srcDir #.encode("latin_1")
     
-    prepare_working_dir(args.outDir)
+    prepare_working_dir(args.outDir, no_del=True)
     
     # -- Find the configuration file
     confFile  = args.confFile or pj(split(__file__)[0], args.project+"_conf.py")        
@@ -571,32 +593,35 @@ def main():
                                 
     # -- if args.srcDir contains "ONLINE_EGGS", this means that we will look for eggs on PYPI and GForge.
     #   - if args.private_packages is True, we add private gforge packages.
+    #   - TODO! Be smart and don't download if already here!
     if args.fetch_online:
-        log.set_verbosity(2)
+        log.set_verbosity(1)
         dldir = pj(args.outDir, "dl_eggs")
+        prepare_working_dir(dldir, no_del=True)
         if args.private_packages:
             add_private_gforge_repositories(args.login, args.passwd)
         for egg, info in dependencies.iteritems():
-            if bt(info[0], EGG):
+            if bt(info[0], EGG) or bt(info[0], ZIPDIST) or bt(info[0], EXEDIST):
                 info[1] = download_egg(egg, dldir)
-                out("Online egg %s downloaded to %s\n"%(egg, info[1])
+                print "Online egg %s downloaded to %s"%(egg, info[1])
 
     # -- find out the installers to package for this mega installer --
-    ok = find_installer_files(args.outDir, args.srcDir, args.pyMaj, args.pyMin, args.arch, 
+    ok = find_installer_files(args.outDir, args.tpp_eggDir, args.srcDir, args.pyMaj, args.pyMin, args.arch, 
                               dependencies)                            
                             
     if not ok:
         sys.exit(-1)
+        
     # -- find out package testing python module names for the packages that need to be tested
-    out("Gathering paths to testing scripts... \n")
+    print "Gathering paths to testing scripts..."
     for pk, info in dependencies.iteritems():
         test    = pj(__path__, pk+"_test.py") if bt(info[0],TEST_ME) else None
         info[2] = test
-        out("\tWill install %s for %s\n"%(test,pk))
+        print "\tWill install %s for %s"%(test,pk)
 
     proj_egg_pths = get_project_eggs(args.arch, args.eggGlobs, args.outDir, args.srcDir)
-    configure_inno_setup(appname, appversion, dependencies, args, funcs, proj_egg_pths)
-    print "Done, please check the generated file."
+    gen = configure_inno_setup(appname, appversion, dependencies, args, funcs, proj_egg_pths)
+    print "Done, please check the generated file:", gen
